feat: entities 정의 완료
