+++ src/entities/index.ts
@@ -1,7 +1,3 @@
 export * from './account';
-export * from './friend';
-export * from './moim';
-export * from './notice';
-export * from './notification';
 export * from './nurse';
 export * from './ward';
