--- src/entities/moim/index.ts
@@ -1,1 +1,0 @@
-export * from './model/queries';
