--- src/entities/notice/index.ts
@@ -1,1 +1,0 @@
-export * from './model/queries';
