--- src/entities/notification/index.ts
@@ -1,1 +1,0 @@
-export * from './model/queries';
