+++ src/features/shift/useCreateShift/index.ts
@@ -3,7 +3,7 @@ import toast from 'react-hot-toast';
 import {match} from 'ts-pattern';
 import useAuth from '@/features/auth/useAuth';
 import useLoading from '@/features/ui/useLoading';
-import {ShiftAPI} from '@/shared/api';
+import {WardAPI} from '@/shared/api';
 import axiosInstance from '@/shared/api/client';
 import useEditShift from '../useEditShift';
 
@@ -102,7 +102,7 @@ function useCreateShift() {
                     });
                 }
 
-                updateShiftPromises.push(ShiftAPI.updateShifts(wardId, wardShiftsDTO));
+                updateShiftPromises.push(WardAPI.updateShifts(wardId, wardShiftsDTO));
             }
 
             await Promise.all(updateShiftPromises);
