+++ src/features/shift/useEditShift/index.ts
@@ -5,8 +5,8 @@ import toast from 'react-hot-toast';
 import {match} from 'ts-pattern';
 import useAuth from '@/features/auth/useAuth';
 import useLoading from '@/features/ui/useLoading';
-import {NurseAPI, ShiftAPI, ShiftTeamAPI, WardAPI} from '@/shared/api';
-import {type WardShiftsDTO} from '@/shared/api/shift/type';
+import {NurseAPI, WardAPI} from '@/shared/api';
+import {type WardShiftsDTO} from '@/shared/api/ward/type';
 import {type Shift} from '@/shared/types/shift';
 import {type WardShiftType, type WardConstraint, type ShiftTeam} from '@/shared/types/ward';
 import {events, sendEvent} from 'analytics';
@@ -42,7 +42,7 @@ const useEditShift = (activeEffect = false) => {
     const {data: shiftTeams} = useQuery({
         queryKey: shiftTeamQueryKey,
         queryFn: async () => {
-            const res = await ShiftTeamAPI.getShiftTeams(wardId!);
+            const res = await WardAPI.getShiftTeams(wardId!);
 
             if (currentShiftTeamId) {
                 if (res.every((x) => x.shiftTeamId !== currentShiftTeamId)) {
@@ -69,7 +69,7 @@ const useEditShift = (activeEffect = false) => {
     const {data: shift, status: shiftStatus} = useQuery({
         queryKey: shiftQueryKey,
         queryFn: async () => {
-            const res = await ShiftAPI.getShift(wardId!, currentShiftTeamId!, year, month);
+            const res = await WardAPI.getShift(wardId!, currentShiftTeamId!, year, month);
 
             if (res === null) return;
 
@@ -87,7 +87,7 @@ const useEditShift = (activeEffect = false) => {
     });
     const {mutate: mutateShift, status: changeStatus} = useMutation({
         mutationFn: ({wardId, focus, shiftTypeId}: {wardId: number; focus: Focus; shiftTypeId: number | null}) =>
-            ShiftAPI.updateShift(wardId, year, month, focus.day + 1, focus.shiftNurseId, shiftTypeId),
+            WardAPI.updateShift(wardId, year, month, focus.day + 1, focus.shiftNurseId, shiftTypeId),
         onMutate: async ({focus, shiftTypeId}) => {
             await queryClient.cancelQueries({queryKey: ['shift']});
 
@@ -154,7 +154,7 @@ const useEditShift = (activeEffect = false) => {
     });
     const {mutate: mutateShiftsAndHistory} = useMutation({
         mutationFn: ({wardId, wardShiftsDTO}: {wardId: number; wardShiftsDTO: WardShiftsDTO; lastFocus: Focus; diff: number}) =>
-            ShiftAPI.updateShifts(wardId, wardShiftsDTO),
+            WardAPI.updateShifts(wardId, wardShiftsDTO),
         onMutate: async ({wardShiftsDTO, lastFocus, diff}) => {
             await queryClient.cancelQueries({queryKey: ['shift']});
 
@@ -229,7 +229,7 @@ const useEditShift = (activeEffect = false) => {
     });
     const {mutate: postShiftMutate, isPending: postShiftLoading} = useMutation({
         mutationFn: ({wardId, shiftTeamId, year, month}: {wardId: number; shiftTeamId: number; year: number; month: number}) =>
-            ShiftAPI.postShift(wardId, shiftTeamId, year, month),
+            WardAPI.postShift(wardId, shiftTeamId, year, month),
     });
     const changeMonth = useCallback(
         (type: 'prev' | 'next') => {
@@ -384,7 +384,7 @@ const useEditShift = (activeEffect = false) => {
                     });
                 }
 
-                updateShiftPromises.push(ShiftAPI.updateShifts(wardId, wardShiftsDTO));
+                updateShiftPromises.push(WardAPI.updateShifts(wardId, wardShiftsDTO));
             }
 
             await Promise.all(updateShiftPromises);
