+++ src/shared/api/auth/index.ts
@@ -4,7 +4,6 @@ import axiosInstance from '../client';
 import {type IAuthAPI} from './type';
 
 class AuthAPI implements IAuthAPI {
-    getAccountMe = async () => (await axiosInstance.get<Account>('/accounts/me')).data;
     demoStart = async () => (await axiosInstance.post<{wardResDto: Ward; accountResDto: Account; accessToken: string}>('/demo/start')).data;
     logout = async (accessToken: string | null) => (await axiosInstance.post('/token/blacklist', {accessToken})).data;
 }
