+++ src/shared/api/index.ts
@@ -2,7 +2,4 @@ export {default as AccountAPI} from './account';
 export {default as AuthAPI} from './auth';
 export {default as FileAPI} from './file';
 export {default as NurseAPI} from './nurse';
-export {default as ShiftAPI} from './shift';
-export {default as ShiftTeamAPI} from './shiftTeam';
-export {default as ShiftTypeAPI} from './shiftType';
 export {default as WardAPI} from './ward';
