+++ src/shared/api/ward/type.ts
@@ -1,7 +1,11 @@
-import {type WaitingNurse} from '@/shared/types/nurse';
-import {type Ward, type WardConstraint, type WardShiftType} from '@/shared/types/ward';
+import {type DutyRequest} from '@/shared/types/request';
+import {type RequestShift, type Shift} from '@/shared/types/shift';
+import {type WaitingNurse, type Nurse} from '@/shared/types/nurse';
+import {type Ward, type WardConstraint, type WardShiftType, type ShiftTeam} from '@/shared/types/ward';
+import {type UpdateNurseDTO} from '../nurse/type';
 
 export interface IWardAPI {
+    // Ward APIs
     // GET
     getWard: (wardId: number) => Promise<Ward>;
     getWardConstraint: (wardId: number, shiftTeamId: number) => Promise<WardConstraint>;
@@ -18,6 +22,57 @@ export interface IWardAPI {
     // DELETE
     deleteWatingNurses: (wardId: number, nurseId: number) => Promise<void>;
     quitWard: (wardId: number) => Promise<void>;
+
+    // Shift APIs
+    // GET
+    getReqShift: (wardId: number, shiftTeamId: number, year: number, month: number) => Promise<RequestShift>;
+    getShift: (wardId: number, shiftTeamId: number, year: number, month: number) => Promise<Shift>;
+    getRequestList: (wardId: number, shiftTeamId: number, year: number, month: number) => Promise<DutyRequest[]>;
+    // PATCH
+    updateShift: (
+        wardId: number,
+        year: number,
+        month: number,
+        day: number,
+        shiftNurseId: number,
+        wardShiftTypeId: number | null,
+    ) => Promise<null>;
+    updateShifts: (wardId: number, wardShifts: WardShiftsDTO) => Promise<void>;
+    updateReqShift: (
+        wardId: number,
+        year: number,
+        month: number,
+        day: number,
+        shiftNurseId: number,
+        wardShiftTypeId: number | null,
+    ) => Promise<void>;
+    acceptRequestShift: (wardId: number, reqShiftId: number, isAccepted: boolean | null) => Promise<void>;
+    // POST
+    postShift: (wardId: number, shiftTeamId: number, year: number, month: number) => Promise<void>;
+
+    // ShiftTeam APIs
+    // GET
+    getShiftTeamNurses: (wardId: number, shiftTeamId: number) => Promise<Nurse[]>;
+    getShiftTeams: (wardId: number) => Promise<ShiftTeam[]>;
+    // POST
+    addNurseIntoShiftTeam: (wardId: number, shiftTeamId: number, addShiftTeamNurseDTO: UpdateNurseDTO) => Promise<Nurse>;
+    createShiftTeam: (wardId: number) => Promise<ShiftTeam>;
+    buildShiftTeam: (wardId: number, shiftTeamId: number, year: number, month: number) => Promise<ShiftTeam>;
+    // PATCH
+    updateShiftTeam: (wardId: number, shiftTeamId: number, updateShiftTeamDTO: UpdateShiftTeamDTO) => Promise<ShiftTeam>;
+    // DELETE
+    removeNurseFromShiftTeam: (wardId: number, shiftTeamId: number, nurseId: number) => Promise<Nurse>;
+    deleteShiftTeam: (wardId: number, shiftTeamId: number) => Promise<ShiftTeam>;
+
+    // ShiftType APIs
+    // GET
+    getShiftTypes: (wardId: number) => Promise<WardShiftType[]>;
+    // POST
+    createShiftType: (wardId: number, createShiftTypeDTO: CreateShiftTypeDTO) => Promise<WardShiftType>;
+    // PUT
+    updateShiftType: (wardId: number, shiftTypeId: number, createShiftTypeDTO: CreateShiftTypeDTO) => Promise<WardShiftType>;
+    // DELETE
+    deleteShiftType: (wardId: number, shiftTypeId: number) => Promise<void>;
 }
 
 export type CreateWardDTO = {
@@ -28,3 +83,16 @@ export type CreateWardDTO = {
 };
 
 export type EditWardDTO = Pick<Ward, 'name' | 'hospitalName'>;
+
+export type WardShiftsDTO = {
+    shiftNurseId: number;
+    date: string;
+    wardShiftTypeId: number | null;
+}[];
+
+export type UpdateShiftTeamDTO = Pick<ShiftTeam, 'name'>;
+
+export type CreateShiftTypeDTO = Pick<
+    WardShiftType,
+    'name' | 'shortName' | 'color' | 'startTime' | 'endTime' | 'isOff' | 'isDefault' | 'isCounted' | 'classification'
+>;
